Add vendor and packager details to the releasenotes. This was something I think was sorely missing before and something that's standard on Solaris.
Make it possible to automatically have modified configure_args put into
relnotes. It's a bit of a hack where instead of assigning the value
directly to the variable we go through a function. Now the nasty bit
is that we can't add relnotes.txt to META_CLEAN since generic_pack
does clean meta which would wipe out changes from build stage.
Instead we add a hack to the distclean method in clean that removes this
file.
I may eventually have to create a generic_clean method that can
be called from pr. platform specific clean function to avoid this kind
of layer violation.

Fix the format of the spec version field again.
I just realised a fairly obvious flaw with the existing system where
increasing the pkgrev from single digits to multiple digits didn't increase
the version but actually decreased it.
Ie. 1.19-9 becomes 11990000 and 1.19-10 becomes 11910000 - ooops!
Instead the pkgrev should be right justified, so instead of 11910000 we'd
have 11900010.
I also eliminated the need for printf in fix_ver, something I shou've
done from the beginning when I realised that Irix 5.3 printf was not up to
the task (only supports %s).